requirements.txt
backports.functools-lru-cache==1.5
cheroot==6.3.1
CherryPy==15.0.0
future==0.16.0
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
portend==2.2
pytz==2018.4
redis==2.10.6
redis-limpyd==1.2
six==1.11.0
tempora==1.11
